+++ .github/workflows/tests.yml
@@ -46,6 +46,25 @@ jobs:
       - name: Test (no default features)
         run: cargo test --no-default-features
 
+  formal-verif:
+    name: Formal Verification
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout sources
+        uses: actions/checkout@v2
+      - name: Install stable toolchain
+        uses: actions-rs/toolchain@v1
+        with:
+          toolchain: stable
+          override: true
+
+      - name: Install Kani
+        run: cargo install --locked kani-verifier && cargo-kani setup
+      
+      - name: Test (no default features)
+        run: cargo test --no-default-features
+
+
   lints:
     name: Lints
     runs-on: ubuntu-latest
